Update Exercise 5-2.c

+++ C Programming Language/CHAPTER 5: Pointers and Arrays/5.2 Pointers and Function Arguments/Exercise 5-2.c
@@ -1,9 +1,9 @@
 #include <ctype.h>
 #include <stdio.h>
-int getfloat(double *pn) 
+int getfloat(float *pn) 
 {
     int c,sign;
-    double power;
+    float power;
     while(isspace(c = getch()));
 
     if(!isdigit(c) && c != EOF && c != '+' && c != '-' && c != '.')
